Fix: 그룹 전체 시간표 조회 API 위치 변경 Refactor: Swagger 변수 설명 추가
#68
